+++ api/python/PE/pyUtils.cpp
@@ -23,15 +23,21 @@ namespace PE {
 
 void init_utils(py::module& m) {
 
+  py::enum_<IMPHASH_MODE>(m, "IMPHASH_MODE",
+      "Enum to define the behavior of :func:`~lief.PE.get_imphash`")
+    .value("DEFAULT", IMPHASH_MODE::DEFAULT, "Default implementation")
+    .value("LIEF",    IMPHASH_MODE::LIEF,    "Same as DEFAULT")
+    .value("PEFILE",  IMPHASH_MODE::PEFILE,  "Use pefile algorithm")
+    .value("VT",      IMPHASH_MODE::VT,      "Same as PEFILE since Virus Total is using pefile");
 
   m.def("is_pe",
       static_cast<bool (*)(const std::string&)>(&is_pe),
-      "Check if the given file is a ``PE``",
+      "Check if the given file is a ``PE``",
       "file"_a);
 
   m.def("is_pe",
       static_cast<bool (*)(const std::vector<uint8_t>&)>(&is_pe),
-      "Check if the given raw data is a ``PE``",
+      "Check if the given raw data is a ``PE``",
       "raw"_a);
 
   m.def("get_type",
@@ -47,19 +53,27 @@ void init_utils(py::module& m) {
 
   m.def("get_imphash",
       &get_imphash,
-      "Compute the hash of imported functions\n\n"
+      R"delim(
+      Compute the hash of imported functions
 
-      "Properties of the hash generated:\n"
-      "\t* Order agnostic\n"
-      "\t* Casse agnostic\n"
-      "\t* Ordinal (**in some extent**) agnostic\n\n"
+      Properties of the hash generated:
 
-      ".. warning::\n\n"
-      "\tThe algorithm used to compute the *imphash* value has some variations compared to Yara, pefile, VT implementation\n"
+        * Order agnostic
+        * Casse agnostic
+        * Ordinal (**in some extent**) agnostic
+        *
 
-      ".. seealso::\n\n"
-      "\thttps://www.fireeye.com/blog/threat-research/2014/01/tracking-malware-import-hashing.html\n",
-      "binary"_a);
+      If one needs the same output as Virus Total (i.e. pefile), you can use :attr:`~lief.PE.IMPHASH_MODE.PEFILE`
+      as second parameter.
+
+      .. warning::
+          The algorithm used to compute the *imphash* value has some variations compared to Yara, pefile,
+          VT implementation
+
+      .. seealso::
+          https://www.fireeye.com/blog/threat-research/2014/01/tracking-malware-import-hashing.html
+      )delim",
+      "binary"_a, "mode"_a = IMPHASH_MODE::DEFAULT);
 
   m.def("resolve_ordinals",
       &resolve_ordinals,
@@ -67,7 +81,7 @@ void init_utils(py::module& m) {
 
       "The ``strict`` boolean parameter enables to throw a " RST_CLASS_REF(lief.not_found) " exception "
       "if the ordinal can't be resolved. Otherwise it skips the entry.",
-      "import"_a, "strict"_a = false,
+      "import"_a, "strict"_a = false, "use_std"_a = false,
       py::return_value_policy::copy);
 }